+++ disk.cpp
@@ -1,12 +1,12 @@
 #include "disk.h"
 
-Disk::Disk(float radius, const Vec3f &position, const Vec3f &normal)
+Disk::Disk(float radius, const Vec3d &position, const Vec3d &normal)
 :Plane(normal), r(radius), rSquared(r*r)
 {
 	pos = position;
 
 	// Infer plane offset from requested position & normal.
-	/*Vec3f offset;
+	/*Vec3d offset;
 	if (n.x != 0)
 		offset.x = pos.x / n.x;
 	if (n.y != 0)
@@ -18,17 +18,17 @@ Disk::Disk(float radius, const Vec3f &position, const Vec3f &normal)
 	d = n.dot(pos);
 }
 
-bool Disk::hit(const Ray &ray, float &t) const
+bool Disk::hit(const Ray &ray, double &t) const
 {
 	// Code is extracted from Plane::hit(...) to avoid too much vfunc calls.
 
-	float dist = (d - ray.origin.dot(n)) / (ray.dir.dot(n));
+	double dist = (d - ray.origin.dot(n)) / (ray.dir.dot(n));
 
 	if(dist > 0 && dist < t)
 	{
-		Vec3f p = ray.origin + ray.dir * dist;
+		Vec3d p = ray.origin + ray.dir * dist;
 		p -= pos;
-		float pSquaredL = p.squaredLength();
+		double pSquaredL = p.squaredLength();
 		if (pSquaredL < rSquared)
 		{
 			t = dist;
